[FEAT] Add support for custom tools in Toolbox
Detailed Changes:
- Added a new `tools` parameter to the `Toolbox` constructor
- The `tools` parameter allows passing a list of custom tool names, in addition to the existing `tool_names` parameter

Impact: This change allows developers to customize the set of tools available in the `Toolbox` class, enabling more flexibility in the developer toolbox.

Additional Notes: The `tools` parameter is meant to be used in addition to the `tool_names` parameter, not as a replacement. Both parameters can be used together to configure the available tools.

Related Issues: [<slug>](https://go/issues/<slug>)